backend/requirements.txt
fastapi
uvicorn
python-dotenv
pytesseract
Pillow
openai
python-multipart
pandas
